UPDATE CarbonFootprint.py - fix secondaryEmissionFactorSource field
This change uses the EmissionFactorDS data model introduced in the previous commit,
along with updating the tests to use a single isolated and consistent CarbonFootprint
fixture, as there was too much separation and changes were required in multiple
places to reflect API schema changes, this has reduced code complexity significantly.
