+++ tests/test_repository/test_product_footprints_repository.py
@@ -1,56 +1,13 @@
 from datetime import datetime, timezone
-from unittest import mock
 
 import pytest
-import sqlalchemy
-import pydantic
 
-#from db.models.product_footprint import ProductFootprint
-from db.repository.product_footprints import create_new_product_footprint, retrieve_product_footprint, \
-    list_product_footprints, count_product_footprints
+from db.repository.product_footprints import (
+    create_new_product_footprint, retrieve_product_footprint, list_product_footprints,
+    count_product_footprints
+)
 from schemas.product_footprint import ProductFootprint
 from schemas.carbon_footprint import CarbonFootprint
-
-
-@pytest.fixture(scope="module")
-def valid_carbon_footprint_data():
-    data = {
-        "declaredUnit": "kilogram",
-        "unitaryProductAmount": 100,
-        "pCfExcludingBiogenic": 10,
-        "pCfIncludingBiogenic": 12,
-        "fossilGhgEmissions": 8,
-        "fossilCarbonContent": 5,
-        "biogenicCarbonContent": 4,
-        "dLucGhgEmissions": 2,
-        "landManagementGhgEmissions": 3,
-        "otherBiogenicGhgEmissions": 1,
-        "iLucGhgEmissions": 2,
-        "biogenicCarbonWithdrawal": -1,
-        "aircraftGhgEmissions": 0.5,
-        "characterizationFactors": "AR6",
-        "crossSectoralStandardsUsed": ["PAS 2050"],
-        "productOrSectorSpecificRules": ["CFS Guidance for XYZ Sector"],
-        "biogenicAccountingMethodology": "PEF",
-        "boundaryProcessesDescription": "Description of boundary processes",
-        "referencePeriodStart": datetime(2023, 1, 1, tzinfo=timezone.utc).isoformat(),
-        "referencePeriodEnd": datetime(2023, 12, 31, tzinfo=timezone.utc).isoformat(),
-        "geographyCountrySubdivision": "AU",
-        "geographyCountry": "AU",
-        "geographyRegionOrSubregion": "Australia and New Zealand",
-        "secondaryEmissionFactorSources": ["Source 1", "Source 2"],
-        "exemptedEmissionsPercent": 2.5,
-        "exemptedEmissionsDescription": "Description of exempted emissions",
-        "packagingEmissionsIncluded": True,
-        "packagingGhgEmissions": 0.5,
-        "allocationRulesDescription": "Description of allocation rules",
-        "uncertaintyAssessmentDescription": "Description of uncertainty assessment",
-        "primaryDataShare": 50,
-        "dqi": {"key1": "value1", "key2": "value2"},
-        "assurance": {"key1": "value1", "key2": "value2"}
-    }
-
-    return data
 
 
 @pytest.fixture(scope="module")
